+++ email/src/email/envelope/flag/notmuch.rs
@@ -1,3 +1,8 @@
+//! Module dedicated to Notmuch email envelope flags.
+//!
+//! This module contains flag-related mapping functions from the
+//! [notmuch] crate types.
+
 use log::{debug, warn};
 use notmuch::Message;