+++ email/src/email/envelope/maildir.rs
@@ -1,4 +1,7 @@
 //! Module dedicated to Maildir email envelopes.
+//!
+//! This module contains envelope-related mapping functions from the
+//! [maildirpp] crate types.
 
 use log::{debug, warn};
 use maildirpp::{MailEntries, MailEntry};
@@ -6,14 +9,14 @@ use rayon::prelude::*;
 
 use crate::{Envelope, Envelopes, Flags, Message};
 
-impl From<MailEntries> for Envelopes {
-    fn from(entries: MailEntries) -> Self {
+impl Envelopes {
+    pub fn from_mdir_entries(entries: MailEntries) -> Self {
         Envelopes::from_iter(
             entries
                 .collect::<Vec<_>>()
                 .into_par_iter()
                 .filter_map(|entry| match entry {
-                    Ok(entry) => Some(Envelope::from(entry)),
+                    Ok(entry) => Some(Envelope::from_mdir_entry(entry)),
                     Err(err) => {
                         warn!("cannot parse maildir entry, skipping it: {err}");
                         debug!("cannot parse maildir entry: {err:?}");
@@ -25,8 +28,8 @@ impl From<MailEntries> for Envelopes {
     }
 }
 
-impl From<MailEntry> for Envelope {
-    fn from(entry: MailEntry) -> Self {
+impl Envelope {
+    pub fn from_mdir_entry(entry: MailEntry) -> Self {
         let msg = Message::from(entry.headers());
         Envelope::from_msg(entry.id(), Flags::from_mdir_entry(&entry), msg)
     }