+++ email/src/email/message/template/mod.rs
@@ -1,8 +1,12 @@
+//! Module dedicated to email message templates.
+//!
+//! A template is a simplified version of an email MIME message, based
+//! on [MML](https://www.gnu.org/software/emacs/manual/html_node/emacs-mime/Composing.html).
+
 mod forward;
 mod new;
 mod reply;
 
-use std::io;
 use thiserror::Error;
 
 #[doc(inline)]
@@ -10,10 +14,8 @@ pub use self::{forward::ForwardTplBuilder, new::NewTplBuilder, reply::ReplyTplBu
 
 #[derive(Debug, Error)]
 pub enum Error {
-    #[error("cannot interpret email as template")]
-    InterpretEmailAsTplError(#[source] pimalaya_email_tpl::tpl::interpreter::Error),
-    #[error("cannot parse raw message")]
-    ParseRawMessageError,
-    #[error("cannot build forward template")]
-    BuildForwardTplError(#[source] io::Error),
+    #[error("cannot interpret message as template")]
+    InterpretMessageAsTemplateError(#[source] pimalaya_email_tpl::tpl::interpreter::Error),
+    #[error("cannot interpret message as thread template")]
+    InterpretMessageAsThreadTemplateError(#[source] pimalaya_email_tpl::tpl::interpreter::Error),
 }