+++ email/CHANGELOG.md
@@ -10,6 +10,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Changed
 
 - Moved `backend::sync` module to `account::sync`, and renamed all associated structures `Backend*` by `Account*`.
+- Replaced `From` and `Into` implementations for flags and envelopes by custom functions.
+- Prefixed email message template builders (new, reply and forward) setters by `with_` to match other builders of the codebase.
 
 ### Removed